backend/requirements.txt
# Web & API Framework (Compatible trio)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Configuration Management & Security
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1

# File Handling
pillow==10.0.1
PyPDF2==3.0.1

# OCR Processing (Tesseract must be installed separately)
pytesseract==0.3.10
pdf2image==1.16.3

# Database: MS SQL Server (Vault Storage)
pyodbc==5.0.1
sqlalchemy==2.0.23

# Database: PostgreSQL + Vector Store (Search & Embeddings)
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/LLM - OpenAI & Google AI
openai==1.3.0
google-generativeai==0.3.0

# Testing Framework
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Code Quality & Type Checking
black==23.12.0
flake8==6.1.0
mypy==1.7.1
types-requests==2.31.0.10

# Utilities
numpy==1.24.3
pandas==2.1.3
requests==2.31.0